requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
Jinja2==3.1.6
Werkzeug==2.3.7
itsdangerous==2.2.0
click==8.2.1

# Utilities
python-dotenv==1.1.1
requests==2.31.0
python-dateutil==2.8.2
PyPDF2==3.0.1
python-docx==0.8.11
six==1.17.0
typing_extensions==4.15.0

# Data / ML / NLP
numpy==1.25.0
Pillow==9.5.0
spacy==3.6.0
thinc==8.1.10
preshed==3.0.10
blis==0.7.11
cymem==2.0.11
catalogue==2.0.10
murmurhash==1.0.13
srsly==2.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5

# Optional helpers
beautifulsoup4==4.12.2
dnspython==2.7.0
email-validator==2.1.0
langcodes==3.5.0
language_data==1.3.0
joblib==1.5.1
marisa-trie==1.3.1
MarkupSafe==3.0.2
regex==2025.7.34
smart-open==6.4.0
soupsieve==2.7
typer==0.9.4
typing-inspection==0.4.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.3.4
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
